Added wiring from addChildren to formik
