+++ src/routes/Registration.tsx
@@ -1,6 +1,6 @@
 import { VStack } from '@chakra-ui/react';
 import { NavBar } from '../components/NavBar';
-import AttendeeInformation from '../components/Registration/AttendeeInformation';
+import AttendeeInformation from '../components/Registration/Pages/AttendeeInformation';
 
 export default function Registration() {
 	return (
